Split pick.spec.ts into pick-tests directory by concern
Break the ~382-line pick.spec.ts into a co-located pick-tests/ directory
to stay under the 300-line test-file soft limit:

- pick-to-firebase.spec.ts: pickToFirebase write-path tests + rankingMode
- firebase-to-pick.spec.ts: firebaseToPick read-path tests + rankingMode
- remove-owner-from-pick-options.spec.ts: removeOwnerFromPickOptions tests
- fixtures.ts: shared date constants and makeFirebasePickPublic fixture

No production code changes; all existing test cases preserved.
